Refactor to have a more unified approach across 'firm' and 'person' (#235)
The common search and index functions are now in packages called 'search' and
'index'. The differences between the database access and indexing have been
factored out to the 'firm' and 'person' packages.

Searching is now configured with different functions to handle the 3 cases:
person only, firm only, or both. Having a "both" option is why I chose to
locate this in 'search' instead of the 'firm'/'person' packages.

Cleaned up the commands so they all have help and the list of available
commands is generated, so we no longer miss any out. In update-alias I removed
the -set flag as it was confusing with it now acting on multiple things (and
probably not very useful), instead I added an -explain flag to match the
cleanup-indices command.

In addition I've added more tests where they were missing and added support for
the 'go test -short' flag, now the majority of tests can be run quickly without
needing docker.