Allow indexing when there are no records
The old query would return `null` for `MIN(id)` and `MAX(id)` if there were no records. This caused a Go error in `.Scan()`, since it's expecting both to be `int`.

This change coalesces null values to 0, so that they are always an int.

Now, if no records exist, it will search for the ID range [0,0], which will return no rows and be quietly skipped over.

Fixes VEGA-1659 #patch