Move indices derivation to search.PrepareQuery* functions
This enables us to provide the indices as part of the request
body, rather than having it hard-coded to the handler which is
invoked.

In the case of index-specific handlers (e.g. person handler),
the indices variable is set to a fixed values.

For searchAll, the index will default to all indices, but can
be overridden using index aliases in the request body.